+++ test/filter-generator/NullFilterTest.cpp
@@ -77,3 +77,13 @@ BOOST_AUTO_TEST_CASE(test_ffmpeg_str)
 
   BOOST_CHECK_EQUAL(ffmpeg, "");
 }
+
+
+BOOST_AUTO_TEST_CASE(test_ffmpeg_audio_str)
+{
+  fg::NullFilter filter;
+
+  std::string ffmpeg(filter.ffmpeg_audio_str());
+
+  BOOST_CHECK_EQUAL(ffmpeg, "");
+}